Added Changes
--- Added Exploder Class Zombie
--- Added Exploder Scripting (WIP)
--- Made Changes to Leaper Script (WIP)
--- Added Exploder Controller
--- Added Exploder Animations

+++ Assets/Scripts/Enemy/Leaper.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEngine.AI;
+using System.Collections;
 
 public class Leaper : Enemy
 {
@@ -118,7 +119,7 @@ public class Leaper : Enemy
         {
             if (hit.CompareTag("Player"))
             {
-                GameManager.instance.playerController.TakeDamage(damageAmount);
+                GameManager.instance.playerController.TakeDamage((int)damageAmount);
             }
         }
     }
